\name{FHtest-package}
\alias{FHtest-package}
\alias{FHtest}
\docType{package}
\title{Tests for right and interval-censored survival data based on the Fleming-Harrington class}
\description{FHtest offers several tests for the comparison of two or more survival curves:

             a) The Fleming-Harrington test for right-censored data based on permutations and on counting processes.

             b) Extension of the Fleming-Harrington test for interval-censored data based on a permutation distribution and on a score vector distribution.}
\details{ \tabular{ll}{
    Package: \tab FHtest\cr
    Type: \tab Package\cr
    Version: \tab 0.85\cr
    Date: \tab 2012-07-26\cr
    License: \tab GPL (>= 2)\cr }
}
\author{ Ramon Oller and Klaus Langohr

Ramon Oller <ramon.oller@uvic.cat> }
\references{
        Oller, R. and Gómez, G. (2012). A generalized Fleming and Harrington's class of tests for interval-censored data. The Canadian Journal of
        Statistics. doi: 10.1002/cjs.11139
}
\keyword{survival}
%\seealso{ %~~ Optional links to other man pages, e.g. ~~ %~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~ %}
%\examples{ %~~ simple examples of the most important functions ~~ %}
