\name{print.FHtesticp}
\alias{print.FHtesticp}
\title{Printing method for \code{FHtesticp} object.}
\description{Printing method for \code{FHtesticp} object.}
\usage{
\method{print}{FHtesticp}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
  \item{x}{An object of type \code{FHtesticp}.}
  \item{digits}{Number of digits for printing.}
  \item{\dots}{Additional arguments.}
}
%\details{
%}

\author{R. Oller and K. Langohr}
%\seealso{
%}
%\examples{
%}

\keyword{internal}
