\name{print.FHtestrcp}
\alias{print.FHtestrcp}
\title{Printing method for \code{FHtestrcp} object.}
\description{Printing method for \code{FHtestrcp} object.}
\usage{
\method{print}{FHtestrcp}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
  \item{x}{An object of type \code{FHtestrcp}.}
  \item{digits}{Number of digits for printing.}
  \item{\dots}{Additional arguments.}
}
%\details{
%}

\author{R. Oller and K. Langohr}
%\seealso{
%}
%\examples{
%}

\keyword{internal}
