\name{FHtest-package}
\alias{FHtest-package}
\alias{FHtest}
\docType{package}
\title{Tests for Right and Interval-Censored Survival Data Based on the Fleming-Harrington Class}
\description{This package offers several tests for the comparison of two or more survival curves:

             a) The Fleming-Harrington test for right-censored data based on permutations and on counting processes.

             b) An extension of the Fleming-Harrington test for interval-censored data based on a permutation distribution and on a score vector distribution.}
\details{ \tabular{ll}{
    Package: \tab FHtest\cr
    Type: \tab Package\cr
    Version: \tab 1.51\cr
    Date: \tab 2023-11-30\cr
    License: \tab GPL (>= 2)\cr }
}
\author{ Ramon Oller and Klaus Langohr

Ramon Oller <ramon.oller@uvic.cat> }
\references{
   Oller, R. and Gómez, G. (2012). A generalized Fleming and Harrington's class of tests for interval-censored data. The Canadian Journal of
        Statistics \bold{40}, 501--516.
        
   Oller, R. and Langohr, K. (2017). FHtest: An {R} Package for the Comparison of Survival Curves with Censored Data. Journal of Statistical 
        Software \bold{81}, 1--25.
}
\keyword{survival}
%\seealso{ %~~ Optional links to other man pages, e.g. ~~ %~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~ %}
%\examples{ %~~ simple examples of the most important functions ~~ %}
